social_media_api/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
